+++ src/error.rs
@@ -5,7 +5,7 @@ use log::SetLoggerError;
 #[derive(Debug)]
 pub enum FtailError {
     SetLoggerError(SetLoggerError),
-    NoDriversError,
+    NoChannelsError,
     IoError(std::io::Error),
     PermissionsError(String),
 }
@@ -16,7 +16,7 @@ impl Display for FtailError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             FtailError::SetLoggerError(e) => write!(f, "Error setting logger: {}", e),
-            FtailError::NoDriversError => write!(f, "No drivers were added to the logger"),
+            FtailError::NoChannelsError => write!(f, "No channels were added to the logger"),
             FtailError::IoError(e) => write!(f, "I/O error: {}", e),
             FtailError::PermissionsError(path) => {
                 write!(f, "The path {} is read-only", path)