+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
+## v0.2.0 (2024-12-22)
+
+- Rename drivers to channels
+
 ## v0.1.2 (2024-09-17)
 
 - Rotate log file if file size exceeds `max_file_size`
 - Add `levels` filter to `Config`
 - Add `targets` filter to `Config`
-- Add `test` driver to be used in tests
+- Add `test` channel to be used in tests
 
 ## v0.1.1 (2024-09-14)